utils/requirements.txt
google-api-python-client==2.37.0
oauth2client==4.1.3
rapidfuzz==2.0.11
environs==9.5.0
loguru==0.6.0
vk_api==11.9.8
pika==1.2.1
peewee
psycopg2
psycopg2-binary
Pillow
youtube_dl
moviepy
TG-AutoPoster